% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_model.R
\name{markov_model}
\alias{markov_model}
\title{The Markov model class}
\description{
Markov model class
}
\examples{
# input data
data(smoking)
smoking_inputs <- do.call(input_parameters$new, smoking)


## ------------------------------------------------
## Method `markov_model$new`
## ------------------------------------------------

smoking_markov_model <- markov_model$new(
  n_states = 2,
  n_cycles = 10,
  cycle_length = 0.5,
  n_samples = 1000,
  n_treatments = 2,
  v_state_names = c("Smoking", "Not smoking"),
  v_treatment_names = c("SoC", "SoC with website"),
  lambda = 20000,
  costs_dr = 0.035,
  qalys_dr = 0.035,
  markov_inputs = smoking_inputs,
  v_init_cohort = c(1, 0)
)

## ------------------------------------------------
## Method `markov_model$generate_input_parameters`
## ------------------------------------------------

smoking_markov_model$generate_input_parameters(n_samples = 10000)

## ------------------------------------------------
## Method `markov_model$generate_transition_matrices`
## ------------------------------------------------

smoking_markov_model$generate_transition_matrices()

## ------------------------------------------------
## Method `markov_model$generate_state_costs`
## ------------------------------------------------

smoking_markov_model$generate_state_costs()

## ------------------------------------------------
## Method `markov_model$generate_state_qalys`
## ------------------------------------------------

smoking_markov_model$generate_state_qalys()

## ------------------------------------------------
## Method `markov_model$generate_markov_trace`
## ------------------------------------------------

smoking_markov_model$generate_markov_trace()

## ------------------------------------------------
## Method `markov_model$generate_costs_qalys`
## ------------------------------------------------

smoking_markov_model$generate_markov_trace()
smoking_markov_model$generate_costs_qalys()

## ------------------------------------------------
## Method `markov_model$generate_results_table`
## ------------------------------------------------

smoking_markov_model$generate_results_table(
  i_reference_treatment = 1,
  v_wtp = 150000,
  currency_symbol = "$"
)

## ------------------------------------------------
## Method `markov_model$export_to_excel`
## ------------------------------------------------

xlsx <- tempfile(fileext = ".xlsx")
smoking_markov_model$export_to_excel(
  xlsx,
  startCol = 5,
  startRow = 8
)
xlsm <- tempfile(fileext = ".xlsm")
smoking_markov_model$export_to_excel(
  xlsm,
  startCol = 5,
  startRow = 8,
  psa_startCol = 5,
  psa_startRow = 1
)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_states}}{Number of states for the Markov model}

\item{\code{n_cycles}}{Number of cycles for the cohort simulation}

\item{\code{cycle_length}}{Length of cycle (in unspecified units that must be consistent with costs and probabilities)}

\item{\code{n_parameters}}{Number of parameters in the model}

\item{\code{n_samples}}{Number of probabilistic samples}

\item{\code{n_treatments}}{Number of treatments, including the reference}

\item{\code{v_state_names}}{Vector of length n_states}

\item{\code{v_treatment_names}}{Vector of length n_treatments}

\item{\code{lambda}}{Willingness-to-pay threshold for net monetary benefit}

\item{\code{costs_dr}}{Discount rate for costs per cycle for cohort simulation (e.g., 5\% is costs_dr = 0.05)}

\item{\code{qalys_dr}}{Discount rate for QALYs or other effects per cycle for cohort simulation (e.g., 5\% is qalys_dr = 0.05)}

\item{\code{markov_inputs}}{An object of R6 class input_parameters with description and (optional) sampled values}

\item{\code{a_transition_matrices}}{Array of (time-homogeneous) random transition matrices with dimensions n_samples by n_states by n_states}

\item{\code{a_state_costs}}{Array of (time-homogeneous) random state costs with dimensions n_treatments by n_samples by n_states}

\item{\code{a_state_qalys}}{Array of (time-homogeneous) random state qalys with dimensions n_treatments by n_samples by n_states}

\item{\code{m_one_off_costs}}{Matrix of random one-off costs with dimensions n_treatments by n_samples}

\item{\code{m_one_off_disutilities}}{Matrix of random one-off costs with dimensions n_treatments by n_samples}

\item{\code{a_cycle_costs}}{Array of estimated cycle costs with dimensions n_cycles by n_treatments by n_samples}

\item{\code{a_cycle_qalys}}{Array of estimated cycle QALYs with dimensions n_cycles by n_treatments by n_samples}

\item{\code{m_total_costs}}{Matrix of estimated discounted total costs with dimensions n_treatments by n_samples}

\item{\code{m_total_qalys}}{Matrix of estimated discounted total QALYs with dimensions n_treatments by n_samples}

\item{\code{m_total_costs_undiscounted}}{Matrix of estimated undiscounted total costs with dimensions n_treatments by n_samples}

\item{\code{m_total_qalys_undiscounted}}{Matrix of estimated undiscounted total QALYs with dimensions n_treatments by n_samples}

\item{\code{a_cohort_array}}{Array of cohort vectors with dimensions n_treatments by n_samples by n_cycles by n_states}

\item{\code{v_init_cohort}}{Initial cohort vector shared across treatments and samples}

\item{\code{df_excel_model_settings}}{Data frame giving settings of model exported to Excel (only updated on calling export_to_excel())}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-markov_model-new}{\code{markov_model$new()}}
\item \href{#method-markov_model-generate_input_parameters}{\code{markov_model$generate_input_parameters()}}
\item \href{#method-markov_model-generate_transition_matrices}{\code{markov_model$generate_transition_matrices()}}
\item \href{#method-markov_model-generate_state_costs}{\code{markov_model$generate_state_costs()}}
\item \href{#method-markov_model-generate_state_qalys}{\code{markov_model$generate_state_qalys()}}
\item \href{#method-markov_model-generate_markov_trace}{\code{markov_model$generate_markov_trace()}}
\item \href{#method-markov_model-generate_costs_qalys}{\code{markov_model$generate_costs_qalys()}}
\item \href{#method-markov_model-generate_results_table}{\code{markov_model$generate_results_table()}}
\item \href{#method-markov_model-export_to_excel}{\code{markov_model$export_to_excel()}}
\item \href{#method-markov_model-clone}{\code{markov_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-new}{}}}
\subsection{Method \code{new()}}{
Initialise the Markov model

Function to use Markov modelling to simulate total costs, QALYs and net benefit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$new(
  n_states,
  n_cycles,
  cycle_length,
  n_samples,
  n_treatments,
  v_state_names,
  v_treatment_names,
  lambda,
  costs_dr,
  qalys_dr,
  markov_inputs,
  v_init_cohort
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_states}}{Number of states for the Markov model}

\item{\code{n_cycles}}{Number of cycles for the cohort simulation}

\item{\code{cycle_length}}{Length of cycle (in unspecified units that must be consistent with costs and probabilities)}

\item{\code{n_samples}}{Number of probabilistic samples}

\item{\code{n_treatments}}{Number of treatments, including the reference}

\item{\code{v_state_names}}{Vector of length n_states}

\item{\code{v_treatment_names}}{Vector of length n_treatments}

\item{\code{lambda}}{Willingness-to-pay threshold for net monetary benefit}

\item{\code{costs_dr}}{Discount rate for costs per cycle for cohort simulation (e.g., 5\% is costs_dr = 0.05)}

\item{\code{qalys_dr}}{Discount rate for QALYs or other effects per cycle for cohort simulation (e.g., 5\% is qalys_dr = 0.05)}

\item{\code{markov_inputs}}{An object of R6 class input_parameters description}

\item{\code{v_init_cohort}}{Vector representing initial proportion in each cohort}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An initialised Markov model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model <- markov_model$new(
  n_states = 2,
  n_cycles = 10,
  cycle_length = 0.5,
  n_samples = 1000,
  n_treatments = 2,
  v_state_names = c("Smoking", "Not smoking"),
  v_treatment_names = c("SoC", "SoC with website"),
  lambda = 20000,
  costs_dr = 0.035,
  qalys_dr = 0.035,
  markov_inputs = smoking_inputs,
  v_init_cohort = c(1, 0)
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-generate_input_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-generate_input_parameters}{}}}
\subsection{Method \code{generate_input_parameters()}}{
Function to generate random parameters for the Markov model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$generate_input_parameters(n_samples = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{Number of parameter samples (default NULL and uses current markov_model number of samples)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model$generate_input_parameters(n_samples = 10000)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-generate_transition_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-generate_transition_matrices}{}}}
\subsection{Method \code{generate_transition_matrices()}}{
Function to fill in the transition matrices array
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$generate_transition_matrices()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model$generate_transition_matrices()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-generate_state_costs"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-generate_state_costs}{}}}
\subsection{Method \code{generate_state_costs()}}{
Function to calculate state costs based on input parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$generate_state_costs()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model$generate_state_costs()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-generate_state_qalys"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-generate_state_qalys}{}}}
\subsection{Method \code{generate_state_qalys()}}{
Function to calculate state qalys based on input parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$generate_state_qalys()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model$generate_state_qalys()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-generate_markov_trace"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-generate_markov_trace}{}}}
\subsection{Method \code{generate_markov_trace()}}{
Function to generate the Markov trace (i.e., values for cohort matrix over time)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$generate_markov_trace()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model$generate_markov_trace()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-generate_costs_qalys"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-generate_costs_qalys}{}}}
\subsection{Method \code{generate_costs_qalys()}}{
Function to generate cycle and total costs and QALYs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$generate_costs_qalys()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model$generate_markov_trace()
smoking_markov_model$generate_costs_qalys()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-generate_results_table"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-generate_results_table}{}}}
\subsection{Method \code{generate_results_table()}}{
Function to summarise cost-effectiveness results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$generate_results_table(
  i_reference_treatment = 1,
  v_wtp = c(20000, 30000),
  currency_symbol = "GBP"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i_reference_treatment}}{Index for reference treatment (default 1)}

\item{\code{v_wtp}}{Vector of willingness-to-pay thresholds at which to calculate net benefits}

\item{\code{currency_symbol}}{String with currency symbol to use for row names}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix with column for each treatment and rows with costs, QALY, ICER and benefit summries
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{smoking_markov_model$generate_results_table(
  i_reference_treatment = 1,
  v_wtp = 150000,
  currency_symbol = "$"
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-export_to_excel"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-export_to_excel}{}}}
\subsection{Method \code{export_to_excel()}}{
Function to export the Markov model to an Excel workbook
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$export_to_excel(
  wb_filename,
  startCol = 5,
  startRow = 8,
  psa_startCol = 6,
  psa_startRow = 1,
  i_reference_treatment = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wb_filename}}{Location of the Excel workbook to which sheets will be added}

\item{\code{startCol}}{Starting column number}

\item{\code{startRow}}{Starting row number}

\item{\code{psa_startCol}}{Location of cells to which numbers of states, treatments, and cycles are added}

\item{\code{psa_startRow}}{Location of cells to which numbers of states, treatments, and cycles are added}

\item{\code{i_reference_treatment}}{Index for reference treatment (default 1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean to indicate if the exported file exists (but does not confirm model export was accurate)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{xlsx <- tempfile(fileext = ".xlsx")
smoking_markov_model$export_to_excel(
  xlsx,
  startCol = 5,
  startRow = 8
)
xlsm <- tempfile(fileext = ".xlsm")
smoking_markov_model$export_to_excel(
  xlsm,
  startCol = 5,
  startRow = 8,
  psa_startCol = 5,
  psa_startRow = 1
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-markov_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-markov_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{markov_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
