% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch_class.R
\name{scratch_class}
\alias{scratch_class}
\title{The Markov model class}
\description{
Markov model class
}
\examples{
markov_smoking <- markov_model$new(n_states = 2, n_cycles = 10, n_samples = 1000, n_treatments = 2, v_state_names = c("Smoking", "Not smoking"), v_treatment_names = c("SoC", "SoC with website"), lambda = 20000, costs_dr = 0.035, qalys_dr = 0.035)


## ------------------------------------------------
## Method `scratch_class$new`
## ------------------------------------------------

markov_smoking <- markov_model$new(n_states = 2, n_cycles = 10, n_samples = 1000, n_treatments = 2, v_state_names = c("Smoking", "Not smoking"), v_treatment_names = c("SoC", "SoC with website"), lambda = 20000, costs_dr = 0.035, qalys_dr = 0.035, markov_inputs = smoking_inputs, v_init_cohort = c(1, 0))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_states}}{Number of states for the Markov model}

\item{\code{n_cycles}}{Number of cycles for the cohort simulation}

\item{\code{n_parameters}}{Number of parameters in the model}

\item{\code{n_samples}}{Number of probabilistic samples}

\item{\code{n_treatments}}{Number of treatments, including the reference}

\item{\code{v_state_names}}{Vector of length n_states}

\item{\code{v_treatment_names}}{Vector of length n_treatments}

\item{\code{lambda}}{Willingness-to-pay threshold for net monetary benefit}

\item{\code{costs_dr}}{Discount rate for costs per cycle for cohort simulation (e.g., 5\% costs_dr = 0.05)}

\item{\code{qalys_dr}}{Discount rate for QALYs or other effects per cycle for cohort simulation (e.g., 5\% is qalys_dr = 0.05)}

\item{\code{markov_inputs}}{An object of R6 class input_parameters with description and (optional) sampled values}

\item{\code{a_transition_matrices}}{Array of (time-homogeneous) random transition matrices with dimensions n_samples by n_states by n_states}

\item{\code{a_cohort_array}}{Array of cohort vectors with dimensions n_treatments x n_samples x n_cycles x n_states}

\item{\code{v_init_cohort}}{Initial cohort vector shared across treatments and samples}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-scratch_class-new}{\code{scratch_class$new()}}
\item \href{#method-scratch_class-clone}{\code{scratch_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scratch_class-new"></a>}}
\if{latex}{\out{\hypertarget{method-scratch_class-new}{}}}
\subsection{Method \code{new()}}{
Initialise the Markov model

Function to use Markov modelling to simulate total costs, QALYs and net benefit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scratch_class$new(
  n_states,
  n_cycles,
  n_samples,
  n_treatments,
  v_state_names,
  v_treatment_names,
  lambda,
  costs_dr,
  qalys_dr,
  markov_inputs,
  v_init_cohort
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_states}}{Number of states for the Markov model}

\item{\code{n_cycles}}{Number of cycles for the cohort simulation}

\item{\code{n_samples}}{Number of probabilistic samples}

\item{\code{n_treatments}}{Number of treatments, including the reference}

\item{\code{v_state_names}}{Vector of length n_states}

\item{\code{v_treatment_names}}{Vector of length n_treatments}

\item{\code{lambda}}{Willingness-to-pay threshold for net monetary benefit}

\item{\code{costs_dr}}{Discount rate for costs per cycle for cohort simulation (e.g., 5\% is costs_dr = 0.05)}

\item{\code{qalys_dr}}{Discount rate for QALYs or other effects per cycle for cohort simulation (e.g., 5\% is qalys_dr = 0.05)}

\item{\code{markov_inputs}}{An object of R6 class input_parameters description}

\item{\code{v_init_cohort}}{Vector representing initial proportion in each cohort}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An initialised Markov model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{markov_smoking <- markov_model$new(n_states = 2, n_cycles = 10, n_samples = 1000, n_treatments = 2, v_state_names = c("Smoking", "Not smoking"), v_treatment_names = c("SoC", "SoC with website"), lambda = 20000, costs_dr = 0.035, qalys_dr = 0.035, markov_inputs = smoking_inputs, v_init_cohort = c(1, 0))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scratch_class-clone"></a>}}
\if{latex}{\out{\hypertarget{method-scratch_class-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scratch_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
