% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_samples}
\alias{summarise_samples}
\title{Summarise MCMC samples}
\usage{
summarise_samples(v_samples, n_digits = 3)
}
\arguments{
\item{v_samples}{Vector of samples from an MCMC or PSA simulation}

\item{n_digits}{Number of significant digits to include}
}
\value{
Summary statistics in Mean (95% CrI) style
}
\description{
Function to calculate mean and 95% credible intervals using MCMC or PSA samples
}
\examples{
summarise_samples(v_samples = c(0.2, 0.3, 0.25, 0.4, 0.3, 0.22), n_digits = 2)
}
